+++ objc2-foundation/src/attributed_string.rs
@@ -19,6 +19,7 @@ object! {
     /// framework contains most of the extension methods.
     ///
     /// See [Apple's documentation](https://developer.apple.com/documentation/foundation/nsattributedstring?language=objc).
+    #[derive(Debug, PartialEq, Eq, Hash)]
     unsafe pub struct NSAttributedString: NSObject;
 }