+++ objc2-foundation/src/value.rs
@@ -15,6 +15,9 @@ use objc2::{msg_send, msg_send_id};
 use super::{NSCopying, NSObject};
 
 object! {
+    // `T: Eq` bound correct to prevent `NSValue<f32>` from being `Eq`
+    // (even though `[NAN isEqual: NAN]` is true in Objective-C).
+    #[derive(Debug, PartialEq, Eq, Hash)]
     unsafe pub struct NSValue<T>: NSObject {
         value: PhantomData<T>,
     }